+++ CHANGELOG.md
@@ -70,3 +70,4 @@
 * Added datetime and guid to finalsave
 * Debug save methods now log full filename to console
 * Added ReservedUIDs option to CoreParams to reserve low UID range for preassignment
+* Implemented dialogue trace (WIP)